added argmin center of japSphere. not working good yet.

+++ src/JapSphere/japSphere.cpp
@@ -8,7 +8,7 @@
 double japSphere::getMinOutOfMesh(Eigen::MatrixXd meshPointsS, RowVector3 p) {
     RowVector3 argMin = meshPointsS.row(0);
     double minVal = calcDistance(argMin, p);
-    int temp;
+    double temp;
     for (int i = 1; i < meshPointsS.rows(); i++) {
         RowVector3 q = meshPointsS.row(i);
         if ((temp = calcDistance(q, p)) < minVal) {
@@ -23,11 +23,10 @@ double japSphere::calcDistance(RowVector3 q, RowVector3 p) {
     return sqrt(pow((q[0] - p[0]), 2) + pow((q[1] - p[1]), 2) + pow((q[2] - p[2]), 2));
 }
 
-RowVector3 japSphere::calculateCenterOfSphere(Eigen::MatrixXd meshPointsS, Eigen::MatrixXd meshPointsInnerS,
-                                              int numberOfRowsInner) {
+RowVector3 japSphere::calculateCenterOfSphere(Eigen::MatrixXd meshPointsS, Eigen::MatrixXd meshPointsInnerS, int numberOfRowsInner) {
     RowVector3 argMax = meshPointsInnerS.row(0);
     double maxVal = getMinOutOfMesh(meshPointsS, argMax);
-    int temp;
+    double temp;
     for (int i = 1; i < numberOfRowsInner; i++) {
         RowVector3 p = meshPointsInnerS.row(i);
         if ((temp = getMinOutOfMesh(meshPointsS, p)) > maxVal) {
